Allow initializing enummer with hash
Similar to the Rails `enum` behavior, allow initializing an `enummer`
with a hash where the values correspond to the indices of the bits that
map to the flag.

This is a more robust and backwards-compatible approach to enums, as the
developer doesn't have to worry about migrating historical data if an
enum value is removed.